Refactor: api 호출 줄이기 위해 코드 수정
